HSEARCH-4303 Execute single-valued association tests against multi-valued associations, too
